requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
WTForms
email-validator
Flask-Bcrypt
python-dotenv
psycopg2-binary  # If using PostgreSQL
# or
mysqlclient==2.1.1  # If using MySQL
